framework/libraries/artie-gpio/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "artie-gpio"
readme = "README.md"
version = "0.0.1"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "artie-util",
]

[project.optional-dependencies]
rpi = [
    "RPi.GPIO",
]
